完成平台保证金展示页面

+++ lib/page/ensure_money/models/ensure_list_item_bean_entity.dart
@@ -10,7 +10,7 @@ class EnsureListItemBeanEntity with JsonConvert<EnsureListItemBeanEntity> {
 
 class EnsureListItemBeanList with JsonConvert<EnsureListItemBeanList> {
   int packagePayId;
-  int payAmount;
+  double payAmount;
   String payTime;
   String shopName;
   int supplyId;
